pyproject.toml
[project]
name = "data-agents"
version = "0.1.0"
description = "AI Agents for Data I/O Demo"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.3",
    "faker>=35.2.0",
    "langchain-openai>=0.3.3",
    "langgraph>=0.2.69",
    "openai>=1.60.2",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "typer>=0.15.1",
]

[tool.poetry]
packages = [
    { include = "lead_agent" },
    { include = "sales_agent" },
    { include = "sales_engineer_agent" }
]

